Moved JLR types and helpers
